feat: 添加玩家种子管理功能，支持购买种子并更新玩家信息

+++ src/main/java/cn/jxufe/model/entity/Seed.java
@@ -63,11 +63,11 @@ public class Seed {
 
     @Column(name = "seed_purchase_price")
     @Schema(description = "种子购买价格")
-    private double seedPurchasePrice;
+    private Long seedPurchasePrice;
 
     @Column(name = "fruit_price_per_unit")
     @Schema(description = "每单位水果出售价格")
-    private double fruitPricePerUnit;
+    private Long fruitPricePerUnit;
 
     @Column(name = "land_requirement")
     @Schema(description = "种植土地需求")
